sqlalchemy

+++ app/schemas.py
@@ -1,12 +1,14 @@
+from decimal import Decimal
+
 from pydantic import BaseModel, Field, field_validator
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=127)
-    amount: float
+    amount: Decimal
     description: str | None = Field(None, max_length=255)
 
     @field_validator('amount')
-    def amount_must_be_positive(cls, v: float) -> float:
+    def amount_must_be_positive(cls, v: Decimal) -> Decimal:
         if v <= 0:
             raise ValueError("Amount must be positive")
         return v
@@ -20,7 +22,7 @@ class OperationRequest(BaseModel):
 
 class CreateWalletRequest(BaseModel):
     name: str = Field(..., max_length=127)
-    initial_balance: float = 0
+    initial_balance: Decimal = 0
 
     @field_validator('name')
     def name_not_empty(cls, v: str) -> str:
@@ -30,7 +32,7 @@ class CreateWalletRequest(BaseModel):
         return v
     
     @field_validator('initial_balance')
-    def initial_balance_must_be_positive(cls, v: float) -> float:
+    def initial_balance_must_be_positive(cls, v: Decimal) -> Decimal:
         if v < 0:
             raise ValueError("initial_balance must be positive")
         return v
